Added some stuff with permission granting and JNI-arrays. We also have added StrictMode to fix things with URI.
There are still problems with this:
WARNING: QApplication was not created in the main() thread
QObject::setParent: Cannot set parent, new parent is in a different
thread
...
QApplication::exec: Must be called from the main thread
Parent is QApplication(0xd477f464), parent's thread is
QThread(0xe2468ca8), current thread is QThread(0xe2468390)
QObject: Cannot create children for a parent that is in a different
thread.
WARNING: QApplication was not created in the main() thread.